TES v0.3 support removed abstaction layers for clarity
