+++ .travis.yml
@@ -18,15 +18,14 @@ install:
   - pip install flake8
 
 before_script:
-  - wget https://github.com/ohsu-comp-bio/funnel/releases/download/0.2.0/funnel-linux-amd64.tar.gz -O /tmp/funnel-linux-amd64.tar.gz
-  - tar -zxvf /tmp/funnel-linux-amd64.tar.gz -C /tmp
-  - mv /tmp/funnel-linux-amd64 /tmp/funnel
-  - export PATH=$PATH:/tmp
+  - wget -O /opt/funnel.tar.gz https://github.com/ohsu-comp-bio/funnel/releases/download/0.4.1/funnel-linux-amd64-0.4.1.tar.gz
+  - tar -zxvf /opt/funnel.tar.gz -C /opt
+  - export PATH=$PATH:/opt
   - docker pull python:2.7
 
 script: 
   - flake8 ./cwl_tes
-  - nosetests -w ./tests
+  - python -m nose ./tests
 
 after_failure:
   - echo "FUNNEL LOGS ----------"
