+++ cwl_tes/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.4"
+__version__ = "0.2.0"
