+++ requirements.txt
@@ -1,4 +1,4 @@
-cwltool==1.0.20170727112954
+cwltool==1.0.20171107133715
 future>=0.16.0
 requests>=2.18.2
-py-tes>=0.1.5
+py-tes>=0.2.0
