+++ tests/requirements.txt
@@ -1,3 +1,4 @@
-cwltest>=1.0.20170214185319
+cwltest>=1.0.20170809112706
+flake8>=3.5.0
 nose>=1.3.7
 PyYAML>=3.12
